requirements.txt
streamlit
requests
cohere
pyrebase4
oauth2client
python-jwt
pycryptodome
